pyproject.toml
[project]
name = "ascii-art"
version = "0.1.0"
description = "Convert any image to ASCII art"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "streamlit>=1.40.1",
]
